[Done] pipeline sinh ra plan v1

+++ src/computation/distance.py
@@ -3,4 +3,4 @@ from src.domain.point import Point
 
 # khoảng các là km
 def calc_km_distance(a: Point, b: Point):
-    return np.sqrt((a.x - b.x) ** 2 + (a.y - b.y) ** 2)
+    return float(np.sqrt((a.x - b.x) ** 2 + (a.y - b.y) ** 2))
